Switch from #if to #ifdef for compile switches
I had once preferred to force the definition of compile-time settings
that had to be chosen to be set explicitly to 1 or 0, because that
makes it easier to check for typos (as `#if SOME_SETTINNG` would notice
that it's not defined).

That practice may make sense for a large project's coding standards,
but it just makes Thinker-Qt harder to compile for those wishing to
experiment with it (I'm not working with it at the moment).  So this
switches to just use #ifdef.