Get rid of the `f` from double initialization constants
The Mandelbrot example explicitly annotates the initialization of
double values with `f` to indicate it's floating point.  This can
trigger a compiler warning that the float is being implicitly expanded
to a double.  While this warning could be suppressed, there seems to
be no good reason for the annotation (despite it existing in the
example sources even today).

This just removes the f.