Move ThinkerBase to separate header file
The "convenience" form of ThinkerManager (which uses a single implicit
global manager) was using a pattern of inclusion from "thinker.h"
which included "thinkermanager.h"...that itself included "thinker.h".
This meant the ThinkerManager was not fully defined.

Something about this apparently used to work (because the include
guard protected against a second inclusion?)  But likely due to changes
in how the MetaObject (moc) compiler works, the moc form of the build
wasn't handling this properly.

This untangles the issue by just moving ThinkerBase into a separate
header file, then the thinkermanager.h includes that thinkerbase.h,
which doesn't have any reference to ThinkerManager in it.