Fix SignalThrottler Qt5 issues, simplify code
The SignalThrottler class exists because there are frequent emit
updates in the thinker code, and it was desirable to make it cheap so
that mutexes did not have to be taken on each unlock().  As a result,
it used the concept that the signal throttler's parent was asserted as
being the thread which did the throttled emit.

This was checked by parenting the SignalThrottler to the thread that
emitThrottled() was expected to be called on.  So that would be the
Thinker thread.  This meant that each time a Thinker was reassigned in
the thread pool, a new SignalThrottler was created and attached to it.

Changes in Qt5 to the rules about QTimer affected this, however.  The
tight loop of the Thinker thread with no event loop was not allowed
to call QTimer::start(), meaning the update strategy did not work.
The rule change basically killed the ability to avoid a mutex, as the
assertions that the same thread was doing all the throttling no longer
were holding.

For the moment, this gets things working by just always using a mutex
in the SignalThrottler.  That makes it a more generic threadsafe class
(as with the signal/slot mechanisms in general) at the cost of finesse
on the performance.  This should be revisited to see if something like
the QAtomic classes could help duck the QMutex lock.